requirements.txt
fastapi
uvicorn
openai-whisper
pydantic[dotenv]
python-multipart
numpy
ffmpeg-python